3 tier distributed
distributed system which emulates the Jysk warehouse was created. The
system can run in different processors concurrently. It assures a good
synchronization, speed, reliability and an incremental growth.
The system is able to keep the information about the pallets that are
kept in the warehouse. Using a database, it is possible to find out the
location of a specific pallet in the warehouse and also how many items
it contains.
Another web server is used for getting different statistics depending on
the boss’ needs. For example, it can tell how many pallets are stored
since the warehouse management server was started the last time.